Cast the len element of a struct iovec to size_t since it's signed on some platforms.

+++ tests/lib/fakewrite.c
@@ -71,7 +71,7 @@ writev(int fd UNUSED, const struct iovec *iov, int iovcnt)
         left = 32;
     total = 0;
     for (i = 0; i < iovcnt && left != 0; i++) {
-        n = (iov[i].iov_len < left) ? iov[i].iov_len : left;
+        n = ((size_t) iov[i].iov_len < left) ? iov[i].iov_len : left;
         memcpy(write_buffer + write_offset, iov[i].iov_base, n);
         write_offset += n;
         total += n;
